Remove ReactDOMProduction-test
There is no need for it now. The only test that was special is moved into ReactDOM-test.
